Remove halo exchanges for 'divergence_3d', which no longer exists as of v5.1
The divergence_3d field was used by the 3-d divergence damping mechanism that
was included in MPAS v5.0. However, in MPAS v5.1, this mechanism was replaced
with one that does not make use of a 'divergence_3d' field. Accordingly,
the call to obtain a pointer to this field and to exchange that field's halo
need to be removed.