Remove hardwired PGI OpenACC flags, activate OpenACC with "OPENACC=true"
The Makefiles in the src/ and src/core_atmosphere/dynamics/ directories had
hard-wired OpenACC flags that were specific to the PGI compiler. These prevented
the code from being compiled with any other compilers, or from being compiled
without OpenACC.

This commit removes hard-wired, compiler-specific flags from lower-level
Makefiles, and instead allows OpenACC flags to be specified in a compiler-dependent
way in the top-level Makefile. Compiling with OpenACC is now accomplished by
adding "OPENACC=true" to the build command.

At present, only OpenACC flags for the PGI compiler are provided, and there
are not checks to ensure that the compiler actually does support OpenACC.