Convert 'role' member of mpas_coupler_type to a function
The 'role' member of the mpas_coupler_type is now a function that can return
a role for a task/thread based on any available information in the type itself.

At the moment, we need to call the role( ) function from an mpas_coupler_type
that is not dereferenced from a domain_type, since the PGI compiler seems to
have problems with this:

Eventually, we should be able to switch code like

    mpas_cpl => domain_ptr % mpas_cpl
    if (mpas_cpl % role( ) == N) then
        ...
    end if

to

    if (domain_ptr % mpas_cpl % role( ) == N) then
        ...
    end if

if issues are resolved in the PGI compiler.