Call MPI_Barrier for intercommunicator in MPAS_coupler_sync
Eventually, the MPAS_coupler_sync routine can be implemented in a way
that is not be susceptible to timeouts and does not use large amounts
of CPU cycles. However, as a first implementation so that all roles
must reach an MPAS_coupler_sync call before proceeding, an MPI_Barrier
is sufficient.