Allow cores to define their own roles
With this commit, the 'role' member function of the MPAS_coupler_type is set
based on the 'component_role' function pointed to in the 'core_type' type.
This allows each core to define roles in its own way. At present, the role of
all MPI tasks for cores besides the atmosphere core are always 0.