Add "role leader" function for each core
When creating MPI intercommunicators to join together two roles, we need some
way to identify the MPI task that will act as the "leader" for each of the two
roles. With this commit, each core defines a component_role_leader function in
its core_type in the same way that it previously defined a component_role.