Introduce initial capability to pair two roles for subsequent communication
This commit adds an initial implementation of a new routine,
MPAS_coupler_create_pairing, that can be used to establish communication between
two "roles". The information needed for this communication is stored in a new
member of the mpas_coupler_type, the mpas_coupler_pairing_type.

The pairing of one role with another creates a uni-directional link that is
specific to an element type (e.g., to cells, edges, or vertices), and requires
as input the equivalent of the indexToCellID, indexToEdgeID, or indexToVertexID
fields.