Add nIndices argument to MPAS_coupler_create_pairing
At present, the method used in the MPAS_coupler_create_pairing routine to
generate communication information between to different roles is not intended
to handle halo elements, but rather, only owned elements. The "nIndices"
argument to this routine is used to specify the number of elements from
the index field that represent owned elements.