Rewrite old accumulation diagnostics with new method
New method entails keeping track of a history of accumulated values, either total since start time or total every 15 minutes. These values are then subtracted or summed to obtain the final total. Optional divider can be provided to generate a time-average output as well. The benefit for this approach is that now we get full hourly and 30-minute diagnostics every 15 minutes that are valid (provided the model has ran for a sufficient duration).