Adding accumulate_radiation to diagnostics
When config_bucket_update is set to 'none', these diagnostics are not naturally accumulated. On the other hand, setting it to update would trigger all radiations to update, which involves a large number of variables and a potentially large penalty on the performance. This new subroutine would only run if the update is set to 'none', thus it should also be compatible, i.e. not double counting, when it is accumulated in the radiation scheme.