Move computation of compactHaloInfo_* arrays into mpas_atm_dynamics_init
The compactHaloInfo_* arrays were previously defined as local variables
in the atm_srk3 routine, which required them to be recomputed by calls to
mpas_dmpar_compact_halo_info at the beginning of each dynamics timestep.

Now, the arrays have been moved to module variables, allowing them to be
computed once before the first call to atm_srk3 and to be freed after
the last call to atm_srk3.