Temp gradient is now on height level
Calculating horizontal temperature gradient on height levels instead of isobaric levels. The original goal of computing on isobaric level was due to a written note about the need to do so, but further examination has led to the possibility of using a slightly different form of the thermal wind relation that relies on GradT being on height level instead.