Enable comm lists to be reused across calls to mpas_dmpar_exch_halo_acc(...)
In order to avoid the overhead of constructing mpas_communication_list
structures for the sendList and recvList in each call to
mpas_dmpar_exch_halo_acc, we now pass pointers to sendList and recvList as
arguments to mpas_dmpar_exch_halo_acc. These pointers must initially be
nullified, and they will be built upon the first call to mpas_dmpar_exch_halo_acc.
Subsequent calls will detect that these pointers are non-null and reuse them.
As a consequence, the sendList and recvList communication lists must be
explicitly destroyed at the end of a simulation.

Note 1: The sendList and recvList structures that are passed to a given call
to mpas_dmpar_exch_halo_acc are specific to the dimensionality of the field
whose halo is being exchanged. We therefore need sendList and recvLists for
six types of fields in the atm_srk3(...) routine at present.

Note 2: At present, the modifications to mpas_dmpar_exch_halo_acc do not
support the specification of specific halo layers to be exchanged (via
the haloLayersIn optional argument).