Fix bug in reuse of halo exchange buffers for 3-d fields
In the mpas_dmpar_exch_halo_3d_real_acc(...) routine, we need to ensure that
the tempRBuffer pointer is valid before initiating the MPI_IRecv. On the first
call to mpas_dmpar_exch_halo_3d_real_acc(...), we allocate tempRBuffer and
point to it with commListPtr % rbuffer. However, on subsequent calls, we no
longer allocate tempRBuffer, nor were we setting the tempRBuffer pointer to
point to the previously allocated commListPtr % rbuffer.