Run LW radiation only on tasks that perform the RADIATION role
Similar to the SW radiation changes, with this commit, only MPI tasks
that are assigned the RADIATION role will call the LW radiation scheme.
Necessary input state for the LW rad scheme is transferred from
the INTEGRATION role tasks, calls to the LW rad code are made, and
the output fields (tendencies) are transferred back to the INTEGRATION
tasks.