Update atm_component_role and atm_component_role_leader to allow for 2 MPI tasks
The logic in the atm_component_role and atm_component_role_leader assumed that
the total number of MPI ranks was twice an even number (so that each of the two
roles would have an even number of ranks). However, it can be helpful during
testing and development to assign just one MPI rank to each role, which
necessitated adjustment to the logic in the functions for defining roles and
role leaders.

Now, the atmosphere core can be run with a total of two MPI ranks, with one
rank assigned to each of two roles.