PIO version compatability check
This commit adds a conditional check in the Makefile to check to see if the
user has the correct version of PIO insatalled against the one which was
specified in the make command and gives an error message and suggestions for
fixes if a mismatch is found.

First, it checks to see if either PIO1 or PIO2 are installed and reports an
error if neither of them are installed. It then checks PIO1 or PIO2 depending
on whether or not `USE_PIO2` is set to `true`.