SST Interval Update Check
This commit adds a pre-run check that dertmines if the user has set the sea
surface update interval if `config_sst_update == true`.

If `config_sst_update == true` and the update interval is `none` the model will
abort its run. Converserly if `config_sst_update == false` and the update
interval is set, a warning will be produced to inform the user that they can
turn on updates by setting `config_sst_update` to true in the namelist file.