Use mminlu rather than config_landuse_data when initializing sea ice
In the physics_init_seaice routine, we were previously using config_landuse_data
to determine the category to be used for ice. However, in cases where
the static file is produced with config_landuse_data = MODIFIED_IGPB_MODIS_NOAH,
but the model initial conditions file is produced with
config_landuse_data = USGS, the resulting sea ice points will have the wrong
category.

The solution to this issue implemented in this commit is to use mminlu rather
than config_landuse_data to determine the ice category. The mminlu value comes
from the input static file (or, in the case where static fields and initial
conditions are being processed all at once, from the config_landuse_data option),
and will therefore be consistent with the land use classification scheme used in
the static file.