Use PHYSICS definition in diagnostics sub-Make and pre-processing of Registry.xml
In order to allow for the use of any pre-processor definitions in the PHYSICS
variable to be used in the pre-processing of the Registry.xml file and in
the build of the diagnostics, the PHYSICS variable is now passed to
the diagnostics sub-Make and added to the pre-processing line for the Registry.xml
file.

This commit also modifies the gen_includes target to depend on the existing
core_reg target to generate the Registry_preprocessed.xml file, rather than
duplicating the recipe from the core_reg target.