Fix incorrect sounding location bug
This commit adds a KD-Tree to core_atmosphere/diagnostics to be used to for
detemining the the nearest sound to specified sounding locations. It replaces
the path-based nearest cell method for determining the cell that a soundings is
within.

This fixes a bug where a sounding would be assigned to the wrong cell in a
concave decomposition.