Fix wrong sounding location bug on concave decomps
This commit fixes a possible bug where a sounding could be assigned to the
wrong cell. It replaces the path-based, nearest cell function with a KD-Tree
datastructure that correctly determines nearest cells.

However, this commit does not fix a bug while running regional MPAS-A where a
sounding location that is specified outside of the simulation region is
assigned incorrectly to a boundry cell.