Discard soundings outside of Limited-Area Region
This commit discards soundings that are outside a limited-area domain, rather
then assigning them to incorrect cells.

Using the in_cell function and the KD-Tree added in the last commit, soundings
are discarded if they are determined to be outside of the region; however, it
does not communicate these actions with the user.