Communicate and report sounding discard
This commit adds logic for determining whether a warning log message should be
produced for a sounding or not.

A point that is determined to not be owned by a task (via in_cell), might still
be within the region in another decomposition. Thus, tasks need to have
some cooperation on determining if a sounding is completely outside of the
limited area region.