Minor bug fixes to MPAS->Albany mesh converter
With these changes, the GLF difference between MALI and standalone Albany is down to
10 digits of precision
* write also the labels (floating/dirichlet) for boundary edges and nodes
* print the mesh and ascii files in double precision.
* Fix a small bug in the MPAS interface. In some very peculiar cases we set the thickness
to 2 or 3 meters. Then we change accordingly the surface elevation in those location to
comply with the floating point condition. However, instead of setting
h = (1-rho_i/rho_o) H, we were setting h = (rho_o/rho_i -1) H. The impact on the GLF
is small, 10e-6 relative error or so, but better to fix it.