Ignore null characters when reading char types
This commit modifies the reading of char types from file so that copying
characters stops when the first null character is reached.  Without this
change, if there are null characters present in a string, they are
copied to the MPAS variable.  This can in turn result in unexpected
behavior, e.g., parsing xtime fails.  Such processing was present for
char1 types, but not for char types.

Also, the logic for parsing char1 types is changed slightly to be
simpler and consistent with the new logic added for char types.