MP out of dyn step 04: dynamics_final_lbc_adjustment now in atm_do_timestep
This is the second part of a two part move. First, source code was bundled
together into a routine, with the call being in the same essential location
as originally. Now the call to the routine has been put where it belongs.

This is part of the process of removing intertwined code inside of the
subroutine atm_srk3, and putting the code into a logical order in
atm_do_timestep.

The only remaining piece will be to remove the call to the microphysics
driver from atm_srk3, then the entire sequential timestep will be inside
of the subroutine atm_do_timestep.

modified:   src/core_atmosphere/dynamics/mpas_atm_time_integration.F
modified:   src/core_atmosphere/mpas_atm_core.F