MP out of dyn step 06: rename subroutine atm_timestep to dynamics_timestep
With the restructuring that is going on inside of atm_do_timestep,
there are calls now to physics, lbc, xtime updates, summary and
also a call to the dynamics. The dynamics call (atm_timestep) will
soon not include a call to the microphysics, so the original
atm_timestep will be doing dynamics ONLY. Therefore, the new name
(dynamics_timestep) more accurately reflects the processing that
is going on in that routine.

modified:   src/core_atmosphere/dynamics/mpas_atm_time_integration.F
modified:   src/core_atmosphere/mpas_atm_core.F