MP out of dyn step 07: add new routine physics_part2 (not used yet)
This new routine, physics_part2, has a call to the microphysics
driver. The new routine, now along side of physics_part1, is not called
from anywhere in this commit. The new code is simply added. The next
commit will remove the original call to the microphysics driver and this
new routine will be used in the correct location in atm_do_timestep. This
is the last step just before we pull the call to microphysics out of the
dynamics subroutine atm_srk3.

modified:   src/core_atmosphere/physics/mpas_atmphys_driver.F