New function "mpas_update_gpu_data_on_host" is added
The function 'mpas_update_gpu_data_on_host' is added in the mpas_atm_time_integration.F.
This function gets the mpas variables using mpas_pool_get_array calls and the
openacc update directives in the function will update the variables on the
host. Both dynamics and physics variables can be updated on to the host using this function.
This function is called at the end of every timestep.