Barotropic subcycling loop compacted for computational efficiency
The barotropic solver is advanced only if both the primary and secondary filter weights at
the next barotropic timestep are non-zero. In other words, we loop over the the barotropic
subcycles till we reach the barotropic timestep after which both the primary and secondary
filter weights are zero (and remain so for all subsequent barotropic timesteps).