Fix to reduce lagged radiation memory footprint on GPU and mapping of lagged radiation ranks
Execution of OpenACC code by radiation role ranks increased the GPU memory footprint
of the code. This change checks the role of the ranks and allow only Integration ranks
to execute any OpenACC directives. Thus reduces the GPU memory footprint. This change also
maps the integration ranks and radiation ranks on each node for efficient execution of the
lagged radiation code.
Note: The values of Number of dynamics ranks and the number of tasks per node are
hardcoded to 4 and 8 respectively. And these options will be moved to namelist file in the
future commits.