Modified logic to determine when FEM needs to be recomputed
The mesh needs to be recomputed non only of the vertices ice mask has changed, but also if
the floating edges mask and the cell ice mask has changed.
Note: in order to compare the clee mask with the previous time steps, we would need the cell mask
to have 2 time levels. Howver, if we set the cell mask to have 2 time levels we get different results.
This needs to be fixed.