Introduce CMake build for init_atmosphere and atmosphere cores
This commit introduces a CMake build capability for the MPAS init_atmosphere and
atmosphere cores, with support for the GNU, Intel, and NVHPC compilers.

When invoking cmake, the following must be defined:
 * MPAS_CORE : Either 'init_atmosphere' or 'atmophere'
 * MPAS_COMPILER : One of 'gnu', 'intel', or 'nvhpc'

If the PIO library is found, MPAS will use PIO; otherwise, SMIOL will be used.

Optionally, MPAS_PRECISION can be specified as 'single' or 'double', with builds
defaulting to double-precision reals.

Acknowledgements: Miles Curry and Maryam Abdi-Oskouei both contributed to the
development of this capability.