Specify name of scalars array in call to advance_scalars(...)
The advance_scalars routine now takes as an argument the name of the scalars
array to be advected. The assumption is that, if the name of the array is XYZ,
then there will exist:

(1) An array in the 'state' pool named XYZ with dimensions
    (num_XYZ, nVertLevels, nCells) and two time levels

(2) A dimension, num_XYZ, in the 'state' pool

(3) An array in the 'tend' pool named XYZ_tend with dimensions
    (num_XYZ, nVertLevels, nCells) and one time level

The scalars arrays can either be var_arrays formed from multiple constituents,
each with dimensions (nVertLevels, nCells), or they can simply be vars with
dimensions (num_???, nVertLevels, nCells).