Fix race condition for ezxml library in parallel builds
Under the right conditions, a race condition in the compilation of the ezxml
library could lead to a failed build. The race condition occurs when the MPAS
build tools (input_gen, parse) try to build the ezxml library for use on
login nodes, while the MPAS framework tries to build the ezxml library for
use on batch nodes.

The fix in this commit resolves this issue by modifying the Makefile for
the ezxml library to compile an object file with a specified name. The MPAS
tools specify an object file name of ezxml_tools.o, while the framework
specifies an object file name of ezxml.o.