Use RKIND kind type parameter for real constants defined in mpas_constants
The real-valued parameters in the mpas_constants module were previously set
to literal values without an explicit kind type parameter. This may lead to
differences in simulation results depending on whether MPAS is compiled with
flags to promote the default real kind (e.g, -fdefault-real-8 with the GNU
Fortran compiler).

For example, if RKIND is defined as 8, the 'gravity' constant defined below
will take on different values, depending on whether MPAS is compiled with,
e.g., -fdefault-real-8 (using the GNU compiler) or not:

    real (kind=RKIND), parameter :: gravity = 9.80616

The correct way to initialize real-valued parameters with kind-type RKIND
is to explicitly specify the kind type parameter for the literal value:

    real (kind=RKIND), parameter :: gravity = 9.80616_RKIND