Add prerequisite tags to all necessary test cases
These test cases depend on other test cases to run properly.
Before this change, the only way to know that is via symlinks
to directories outside of this test case.  With this change, it
will be easier for regression test suites to make sure prerequisites
are present and that they always run before the test cases that
depend on them (even if we run tests cases in parallel).