Fix GWDO check on all water regional simulations
Originally, the GWDO check ensured that the GWDO fields were valid by checking
to see if the `var2d` field was not all zero. For global simulations, an
all-zero `var2d` field (with `config_gwdo_scheme` not set to "off") most likely
meant that the user did not interpolate the GWDO fields.

However, regional MPAS simulations that were entirely over water would make
`var2d` and other GWDO fields all zero, which would incorrectly trigger this
error.

This commit updates the GWDO check to only be thrown if any land is present in
the simulation.