Fix typos in Registry descriptions of rthratenlw and rthratensw
This commit fixes the description of rthratenlw which previously stated that it
was due to 'short wave' radiation instead of 'long wave' radiation.

The descriptions for rthratenlw and rthratensw reference 'long wave' and
'short wave', respectfully, as two words. Other descriptions in the registry
reference these as 'longwave' and 'shortwave', i.e. as a single word, so this
commit changes 'long wave' and 'short wave' to a single word to match the rest
of the registry. This helps when finding fields that are related to shortwave
and longwave radiation.