Add missing initialization of some scalars field members in atm_allocate_scalars
In the atm_allocate_scalars routine, a field3dReal type is allocated for both
scalars and scalars_tend. This commit adds initialization of the prev, next,
sendList, recvList, and copyList members of these instances to avoid potential
problems later on in the model start-up with the use of uninitialized pointers.