Remove -m64 from all compiler flag sets in gfortran build target
The -m64 option is an architecture-specific option, available for the PowerPC
and x68 architectures, but not for the AArch64 architecture.

Building with the 'gfortran' target on AArch64 systems results in the following
error:

  gcc: error: unrecognized command line option ‘-m64’

In principle, the -m64 option may be needed on some older PowerPC or x86
machines, though no current examples of systems with this requirement are
known. An alternative to modifying the existing 'gfortran' target would be to
create a new target, 'gfortran-arm' or 'gfortran-aarch64'; however, having
a single 'gfortran' target that is architecture-agnostic seems like a cleaner
solution.

Note: the 'gfortran-clang' and 'gnu-nersc' targets have not been modified to
remove the -m64 flag. These build targets are assumed to have been added for
specific platforms (e.g., NERSC machines and macOS machines with the clang C/C++
compilers).