Add unified static mapping function for static interpolations
This commit adds a subroutine that preforms mapping of a static dataset. Given
two function references, the mapping function has been generalized to map
static datasets to MPAS cells.

As of this commit, that mapping function is not used and thus
mpas_init_atm_static still does not correctly preform static interpolation.