Update shared abstract interfaces to take in geotile_manager
This commit updates the abstract interfaces, interp_criteria_function and
interp_accumulation_function, to take in an instance of the
mpas_geotile_mgr_type. This is needed for the maximum snow albedo interpolation
which needs access to the missing_value variable within mgr % pool.

This commit also updates the instances of these functions
(continuous_interp_criteria, terrain_interp_accumulation, etc.).