HACK: Make N=mu*N to allow N-based runs
This hack sets N=mu*N before passing N to Albany.  This is hack to get
Albany to only evaluate mu*N at cell centers rather than at all
quadrature points.  This hack is used for doing a forward run where mu
is calculated from an MPAS N field.  In this configuration, the albany
input file should mu be spatially uniform with a value of 1.0.

This commit should be removed or formalized before this branch is
merged!!!