Get correct ISO 8601 decimal fractions for date-time strings with external ESMF
The mpas_get_time routine previously obtained date-time strings from
the external ESMF library with the timeString argument to ESMF_TimeGet. However,
for compatibility with the date-time strings returned by the built-in ESMF
time manager library, full ISO 8601 date-time strings should be obtained with
the timeStringISOFrac optional argument to ESMF_TimeGet.

For example, before this commit, when the MPAS infrastructure was compiled with
the MPAS_EXTERNAL_ESMF_LIB macro defined, the following date-time string might
be returned

   2010-10-23_00:00:12:1/2

while the built-in ESMF time manager would return

   2010-10-23_00:00:12.500000000

This commit leads to the latter date-time string being returned regardless of
whether MPAS is using the built-in ESMF time manager or an external ESMF
library.