Remove warnings across unified static interp changes
This commit removes the warnings that are present when compiling with -Wall
with the GNU compiler for the mpas_init_atm_static.F file.

There are still four warnings produced when compiling with -Wall and are in
present because the mgr dummy argument in the calls to some
interp_criteria_functions or interp_accumulation_functions are not used in some
implementations.