Land Ice: Modify tests to comply with the bc refactor in Albany
Following https://github.com/SNLComputation/Albany/pull/696,
linear sliding conditions are implemented as a Power Law condition,
where we set Power Exponent = 1 and Effective Pressure = 1 and muFriciton = beta.

We also link to updated meshes for the Greenland and the Thwaites test cases
that have the fileds muFriction and Effective Pressure.