Enable more flexibility in PIO namelist settings
This commit adds flexibility in how I/O tasks are specified using the namelist
options config_pio_num_iotasks and config_pio_stride. Previously, it was
possible to set the number of I/O tasks to zero, in which case the number of
I/O tasks was set to the total number of MPI ranks (regardless of the setting
of config_pio_stride).

Now, if config_pio_num_iotasks is zero, the number of I/O tasks is set to the
total number of MPI ranks divided by the specified value of config_pio_stride.
Additionally, if the I/O stride is zero, the stride is set to the total number
of MPI ranks dividied by the specified number of I/O tasks.

Also included in this commit are log messages to inform users of the computed
I/O task configuration that will actually be used in a run; for example:

 ----- I/O task configuration: -----

     I/O task count  = 1
     I/O task stride = 72