Add a generic function to map terrestrial datasets to MPAS meshes
This commit adds in a generic function which can be used map terrestrial
datasets to MPAS meshes. The function refactors and abstracts the algorithm
that was first introduced in 8e3859, which enables terrestrial datasets to be
safely parallelized without the need for a CVT partition file.

It also adds in two abstract routines which will need to be implemented by any
new dataset interpolation which wishes to use the unified function.

At this time, the unified function is not used and no implementation of the
abstract routines are created, because of this, a warning is generated for the
mpas_init_atm_static module.

Finally, it also makes `nBdyLayers` a private, module level variable so that it
can be used by both init_atm_static and init_atm_map_static_data.