Add data movement directives to the cloud-fraction scheme driver
This commit adds data movement directives for the cloud-fraction scheme driver.

The updates of the areaCell, meshDensity, and xland fields are not strictly
necessary, but they are added here as a precaution, while the update of the
cldfrac field is required for correctness. A future clean-up commit will likely
remove unneeded updates of areaCell, meshDensity, and xland.