Add comment explaining default initialization choice for att_list_type % attType
The attType member of the att_list_type derived type was given a default value
of -1, but there was no explanation in the code as to how this default value was
chosen. This commit adds a comment explaining that the key consideration in the
choice of the default value for attType is that it must not match a valid type
(e.g., MPAS_ATT_INT, MPAS_ATT_REAL, etc.) so that new instances of the
att_list_type that haven't been explicitly initialized or otherwise set cannot
be mistaken for valid attributes.