If MPAS-Atmosphere is run as a CAM dycore, don't zero-out physics tendencies
Previously, if the DO_PHYSICS macro was not defined, we assumed that
MPAS-Atmosphere was running without physics, in which case we set the physics
tendencies tend_ru_physics, tend_rtheta_physics, and tend_rho_physics to zero.

However, if MPAS-Atmosphere is running as a CAM dycore, the DO_PHYSICS macro
is not defined (since we don't want to activate any of the stand-alone MPAS-
Atmosphere physics), but we do not want to clear the physics tendency fields,
which are set in the CAM-MPAS dynamics-physics coupling interface.

If MPAS-Atmosphere is running as a CAM dycore, the MPAS_CAM_DYCORE will be
defined, so we now only clear physics tendencies if neither DO_PHYSICS nor
MPAS_CAM_DYCORE are defined.