Stop with a critical error if input stream is missing dims for any of its fields
During the second phase of bootstrapping, if a file associated with an input
stream does not define all of the dimensions required by fields to be read from
the stream, stop with a critical error.

Note that fields that are deactivated by packages will not trigger the error
condition added in this commit, since deactivated fields are excluded from
streams by logic in the MPAS_stream_mgr_add_field routine.