Define cpp macro MPAS_BUILD_TARGET as the name of the build target (e.g., llvm)
This commit adds the definition of a new macro, MPAS_BUILD_TARGET, to the
CPPFLAGS variable in the top-level Makefile. The macro is set to the name of
the build target that was used to compile MPAS, for example, 'ifort' or 'llvm'.

Having access to the name of the build target via the MPAS_BUILD_TARGET macro
should (through further code changes) allow MPAS cores to be aware of or to
report the target that was used to build them.