Report in log file the total memory allocated for fields by the block creator
This commit adds code to the mpas_block_creator module to track and
report the total number of bytes allocated for fields by the
mpas_block_creator_allocate_pool_fields routine. Allocated bytes are
reported for both the MPI task itself as well as the sum across all MPI
tasks. The report of the number of allocated bytes is written to the log
file, for example:

 Allocating fields ...
  284 MB allocated for fields on this task
  3408 MB total allocated for fields across all tasks
  ----- done allocating fields -----