Add checks on return status for each field allocation
This commit adds code to obtain a status code (with the optional STAT argument
to the ALLOCATE intrinsic) for each allocation of a block field in the
mpas_block_creator_allocate_pool_fields routine. If a status code is non-zero,
an error message is printed, and a critical error message is generated after the
call to mpas_block_creator_allocate_pool_fields in
mpas_block_creator_finalize_block_phase2 .

Based on some testing, the generation of non-zero statuses from allocations that
should be failing doesn't appear to be entirely reliable. However, checking the
allocation status shouldn't hurt, and even if it catches failures only under
some scenarios, that should be an improvement over the lack of any error
checking until now.