Avoid use of hardwired Fortran unit number 22 in atm_core module
The atm_core module used a hardwired value of 22 for the Fortran unit
number used to read and write the 'restart_timestamp' file. Although not known
to be an issue at present, the use of hardwired unit numbers could in principle
lead to the use of a unit number that was already in use by other code.

To avoid potential problems, the atm_core module now uses mpas_new_unit and
mpas_release_unit from the mpas_io_units module to more safely obtain and
release Fortran unit numbers.