Report runtime settings for MPI and OpenMP in [init_]atmosphere log files
This commit adds code to the setup_log routines for the init_atmosphere and
atmosphere cores to write information about the runtime values of MPI task count
and maximum OpenMP thread count to log files. For example:

 Run-time settings:
   MPI task count: 6
   OpenMP max threads: 4

This information can be helpful in determining the parallel runtime settings
that were used for a run when only log files are available.

Note that, if OPENMP=true was not specified in the build options, then no
reporting of the maximum OpenMP thread count is performed. This allows for
disambiguation of cases where the model was run with a single OpenMP thread or
with no OpenMP support at all.