Change if logic and add print statements for mpas_init FileParam arguments
This commit changes the if tests for namelistFileParam and streamsFileParam so
that the length of an optional parameter is only checked after it is certain
that the parameter has been provided. Some write statments were also provided
with error and warning information. The write statments had to be used since the
logging infrastructure hasn't been enabled until mpas_init_framework_phase1 gets
called.