Add 3-D grid analysis nudging FDDA to MPAS-A
This EPA_FDDA commit will add 3-D grid analysis nudging FDDA to MPAS-A
as an option to assist in keeping retrospective simulations dynamically
on track to improve evaluation against observations.  This FDDA
development for MPAS-A (Bullock et al., 2018) nudges temperature,
humidity, and wind toward target values and is based on the comparable
FDDA feature available in WRF, including MPAS-A versions of many of the
WRF FDDA options. Enabling analysis nudging FDDA in MPAS-A maintains
high fidelity to the reference fields while still conserving mass.  An
extra step (additional details are provided in the associated pull
request) using init_atmosphere_model is necessary to generate the FDDA
file required for input to a subsequent MPAS-A simulation with FDDA
enabled.

New file:
  src/core_atmosphere/physics/mpas_atmphys_fdda.F

Modified files:
  Makefile
  src/core_atmosphere/Registry.xml
  src/core_atmosphere/physics/Makefile
  src/core_atmosphere/physics/mpas_atmphys_control.F
  src/core_atmosphere/physics/mpas_atmphys_driver.F
  src/core_atmosphere/physics/mpas_atmphys_manager.F
  src/core_atmosphere/physics/mpas_atmphys_todynamics.F
  src/core_init_atmosphere/Registry.xml

These EPA_FDDA code changes to MPAS-A are based on the 22 August 2022
"develop" branch of MPAS v7.3.

Reference:
Bullock Jr., O. R., H. Foroutan, R. C. Gilliam, and J. A. Herwehe, 2018:
  Adding four-dimensional data assimilation by analysis nudging to the
  Model for Prediction Across Scales  Atmosphere (version 4.0). Geosci.
  Model Dev., 11, 28972922.  https://doi.org/10.5194/gmd-11-2897-2018